Ermeks

+++ src/main/java/com/sedbproj/Railways/Repositories/EmployeeRepository.java
@@ -3,5 +3,5 @@ package com.sedbproj.Railways.Repositories;
 import com.sedbproj.Railways.Entities.EmployeeEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EmployeeRepository extends CrudRepository<EmployeeEntity, Long> {
+public interface EmployeeRepository extends CrudRepository<EmployeeEntity, Integer> {
 }
